requirements.txt
requests>=2.28.0
boto3>=1.24.0
PyYAML>=6.0
python-dotenv>=0.19.0